+++ src/lib/fetch_cache.json.d.ts
@@ -1,0 +1,5 @@
+declare module '$lib/packages.json' {
+	import type {FetchCache} from '@ryanatkn/orc/fetch_cache.js';
+	const data: FetchCache;
+	export default data;
+}