+++ src/lib/github.ts
@@ -2,8 +2,9 @@ import type {Logger} from '@grogarden/util/log.js';
 import type {PackageMeta} from '@fuz.dev/fuz_library/package_meta.js';
 import {request} from '@octokit/request';
 import {z} from 'zod';
+import {to_fetch_cache_key, type FetchCache, type FetchCacheItem} from './fetch_cache';
 
-export const GithubIssue = z.object({
+export const GithubPullRequest = z.object({
 	url: z.string(),
 	id: z.number(),
 	number: z.number(),
@@ -20,22 +21,33 @@ export const GithubIssue = z.object({
 	updated_at: z.string(),
 	draft: z.boolean(),
 });
-export type GithubIssue = z.infer<typeof GithubIssue>;
+export type GithubPullRequest = z.infer<typeof GithubPullRequest>;
 
-export const fetch_github_issues = async (
+export const fetch_github_pull_requests = async (
 	url: string,
 	pkg: PackageMeta,
+	cache?: FetchCache,
 	log?: Logger,
 	token?: string,
-): Promise<GithubIssue[] | null> => {
+): Promise<FetchCacheItem<GithubPullRequest[] | null>> => {
 	log?.info('url', url);
-	if (!pkg.owner_name) return null;
+	if (!pkg.owner_name) throw Error('owner_name is required');
+	const params = {owner: pkg.owner_name, repo: pkg.repo_name, sort: 'updated'} as const;
+	const key = to_fetch_cache_key(url, params);
+	if (cache?.has(key)) {
+		return cache.get(key)!;
+	}
 	const res = await request('GET /repos/{owner}/{repo}/pulls', {
 		headers: {authorization: 'token ' + token},
-		owner: pkg.owner_name,
-		repo: pkg.repo_name,
-		sort: 'updated',
+		...params,
 	});
 	log?.info(`res`, res);
-	return res.data.map((i) => GithubIssue.parse(i));
+	console.log(`res.header`, res.headers);
+	return {
+		url,
+		params,
+		key,
+		etag: res.headers.etag ?? null,
+		data: res.data.map((i) => GithubPullRequest.parse(i)),
+	};
 };