+++ src/lib/packages.task.ts
@@ -1,7 +1,7 @@
 import type {Task} from '@grogarden/gro';
 import {load_package_json} from '@grogarden/gro/package_json.js';
 import {z} from 'zod';
-import {writeFile} from 'node:fs/promises';
+import {readFile, writeFile} from 'node:fs/promises';
 import {format_file} from '@grogarden/gro/format_file.js';
 import {exists} from '@grogarden/gro/exists.js';
 import {join} from 'node:path';
@@ -10,6 +10,7 @@ import {GITHUB_TOKEN_SECRET} from '$env/static/private';
 
 import {fetch_packages, type MaybeFetchedPackage} from '$lib/fetch_packages.js';
 import {load_orc_config} from '$lib/config.js';
+import {deserialize_cache, serialize_cache, type FetchCache} from '$lib/fetch_cache.js';
 
 // TODO etags - cache?
 // TODO refactor for reusability
@@ -22,9 +23,6 @@ export const Args = z
 				description: 'path to the directory containing the source package.json and orc.config.ts',
 			})
 			.optional(),
-		outfile: z
-			.string({description: 'path to the generated packages.json file'})
-			.default(join(paths.lib, 'packages.json')),
 	})
 	.strict();
 export type Args = z.infer<typeof Args>;
@@ -36,12 +34,19 @@ export const task: Task<Args> = {
 	Args,
 	summary: 'download metadata for the given packages',
 	run: async ({args, log}) => {
-		const {dir, outfile} = args;
+		const {dir} = args;
+
+		const outfile = join(paths.lib, 'packages.json');
+		const cache_path = join(paths.lib, 'fetch_cache.json');
 
 		const orc_config = await load_orc_config(dir);
 		const {packages} = orc_config;
 
-		const fetched_packages = await fetch_packages(packages, GITHUB_TOKEN_SECRET, log);
+		const cache: FetchCache = (await exists(cache_path))
+			? deserialize_cache(await readFile(cache_path, 'utf8'))
+			: new Map();
+
+		const fetched_packages = await fetch_packages(packages, GITHUB_TOKEN_SECRET, cache, log);
 
 		const local_package_json = await load_package_json(dir);
 
@@ -66,6 +71,21 @@ export const task: Task<Args> = {
 				`declare module '$lib/packages.json' {
 	import type {FetchedPackage} from '@ryanatkn/orc/fetch_packages.js';
 	const data: FetchedPackage[];
+	export default data;
+}
+`,
+			);
+		}
+
+		await writeFile(cache_path, await format_file(serialize_cache(cache), {filepath: cache_path}));
+
+		const types_cache_path = cache_path + '.d.ts';
+		if (!(await exists(types_cache_path))) {
+			await writeFile(
+				types_cache_path,
+				`declare module '$lib/fetch_cache.json' {
+	import type {FetchCache} from '@ryanatkn/orc/fetch_cache.js';
+	const data: FetchCache;
 	export default data;
 }
 `,